pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "Matheus Camargo",email = "matheus.crodrigues@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.4,<9.0.0)",
    "factory-boy (>=3.3.1,<4.0.0)",
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "pytest-django (>=4.9.0,<5.0.0)",
    "faker (>=35.0.0,<36.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
